Create IntegrationMethods.cpp
Two methods of approximate integration of functions - the method of rectangles and the method of trapezoids.

The program receives as input the method of integration, the ends of the segment and the number of segments of the partition, and calculates the integral of the functions sin and cos.

EXAMPLE:
Input: rectangle 0 3.1415926 100
Output: 5.3592e-08
2.00008